update to version 1.4.5
git-svn-id: https://plugins.svn.wordpress.org/duplicator/trunk@2708869 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ classes/class.password.php
@@ -163,7 +163,7 @@ class DUP_PasswordHash
 		$itoa64 = './ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
 
 		$output = '$2a$';
-		$output .= chr(ord('0') + $this->iteration_count_log2 / 10);
+		$output .= chr(ord('0') + (int)($this->iteration_count_log2 / 10));
 		$output .= chr(ord('0') + $this->iteration_count_log2 % 10);
 		$output .= '$';
